KERN-1939  Added term vectors for the "readers" field.
This field was being used in the MoreLikeThis query for related groups and, without term vectors, Solr was forced to perform tokenisation on the stored value of the readers field at runtime.

Note that there is a corresponding commit against Nakamura to force MoreLikeThis to ignore fields that don't have term vectors in its similarity calculation.